@@ -32,7 +32,7 @@ class Client
     const SERVER_URI = 'https://gcm-http.googleapis.com/gcm/send';
 
     /**
-     * @var Zend\Http\Client
+     * @var \Zend\Http\Client
      */
     protected $httpClient;
 
@@ -56,7 +56,7 @@ class Client
      *
      * @param string $apiKey
      * @return Client
-     * @throws InvalidArgumentException
+     * @throws Exception\InvalidArgumentException
      */
     public function setApiKey($apiKey)
     {
@@ -70,13 +70,15 @@ class Client
     /**
      * Get HTTP Client
      *
-     * @return Zend\Http\Client
+     * @throws \Zend\Http\Client\Exception\InvalidArgumentException
+     *
+     * @return \Zend\Http\Client
      */
     public function getHttpClient()
     {
         if (!$this->httpClient) {
             $this->httpClient = new HttpClient();
-            $this->httpClient->setOptions(array('strictredirects' => true));
+            $this->httpClient->setOptions(['strictredirects' => true]);
         }
         return $this->httpClient;
     }
@@ -84,7 +86,7 @@ class Client
     /**
      * Set HTTP Client
      *
-     * @param Zend\Http\Client
+     * @param \Zend\Http\Client
      * @return Client
      */
     public function setHttpClient(HttpClient $http)
@@ -96,9 +98,17 @@ class Client
     /**
      * Send Message
      *
-     * @param Mesage $message
+     * @param Message $message
+     *
+     * @throws \Zend\Json\Exception\RuntimeException
+     * @throws \ZendService\Google\Exception\RuntimeException
+     * @throws \Zend\Http\Exception\RuntimeException
+     * @throws \Zend\Http\Client\Exception\RuntimeException
+     * @throws \Zend\Http\Exception\InvalidArgumentException
+     * @throws \Zend\Http\Client\Exception\InvalidArgumentException
+     * @throws \ZendService\Google\Exception\InvalidArgumentException
+     *
      * @return Response
-     * @throws Exception\RuntimeException
      */
     public function send(Message $message)
     {
